+++ fs/partitions/check.h
@@ -2,7 +2,7 @@
  * add_partition adds a partitions details to the devices partition
  * description.
  */
-void add_gd_partition(struct gendisk *hd, int minor, int start, int size);
+void add_gd_partition(struct gendisk *hd, int minor, unsigned long start, unsigned long size);
 
 typedef struct {struct page *v;} Sector;
